Move stats printing to util module
